refactor(watt): update table storybook
